src/sepsidae_site/conf/settings/production.py
from base import *  # noqa

EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
EMAIL_HOST = environ.get('EMAIL_HOST', 'smtp.gmail.com')
EMAIL_HOST_PASSWORD = environ.get('EMAIL_HOST_PASSWORD', '')
EMAIL_HOST_USER = environ.get('EMAIL_HOST_USER', 'your_email@example.com')
EMAIL_PORT = environ.get('EMAIL_PORT', 587)
EMAIL_SUBJECT_PREFIX = '[%s] ' % SITE_NAME
EMAIL_USE_TLS = True
SERVER_EMAIL = EMAIL_HOST_USER

DATABASES['default']['NAME'] = 'sepsidae_site_production'

MIDDLEWARE_CLASSES += (
    'johnny.middleware.LocalStoreClearMiddleware',
    'johnny.middleware.QueryCacheMiddleware',
)

CACHES = {
    'default': {
        'BACKEND': 'johnny.backends.memcached.MemcachedCache',
        'LOCATION': 'mediapop.moaich.cfg.apse1.cache.amazonaws.com:11211',
        'KEY_PREFIX': 'sepsidae_site-production',
        'JOHNNY_CACHE': True,
    }
}

from johnny.cache import enable
enable()

SECRET_KEY = 'a secret key for production'

RAVEN_CONFIG['site'] = 'production'

DEFAULT_FILE_STORAGE = 'storages.backends.s3boto.S3BotoStorage'
THUMBNAIL_DEFAULT_STORAGE = DEFAULT_FILE_STORAGE
AWS_STORAGE_BUCKET_NAME = 'sepsidae_site-production'