requirements.txt
pip==1.5
setuptools==2.1
boto==2.23.0
Fabric==1.8.1

Django==1.6.1
gunicorn==18.0
MySQL-python==1.2.5
Pillow==2.3
South==0.8.4
django-braces==1.3.1
django-model-utils==2.0
logutils==0.3.3
django-storages==1.1.8
raven==4.0.3
django-haystack==2.1.0
pytz>2013b
jack-bower==0.1.7
django-compressor==1.3
johnny-cache==1.4
python-memcached==1.53
coverage==3.7.1
django-debug-toolbar==1.0.1
factory-boy==2.3.0
mock==1.0.1
Sphinx==1.2.1
flake8==2.1.0
easy-thumbnails==1.4